+++ CapstoneProject/Assets/CapstoneProject/Scripts/ConsoleScripts/XMLVendorReader.cs
@@ -33,8 +33,8 @@ public class XMLVendorReader : MonoBehaviour {
 			manager.allWeapons[i].range = float.Parse(firstNode.Attributes.GetNamedItem("range").Value);
 			manager.allWeapons[i].fireRate = float.Parse(firstNode.Attributes.GetNamedItem("fireRate").Value);
 			manager.allWeapons[i].force = float.Parse(firstNode.Attributes.GetNamedItem("force").Value);
-			manager.allWeapons[i].bulletsPerClip = int.Parse(firstNode.Attributes.GetNamedItem("bulletsPerClip").Value);
-			manager.allWeapons[i].maxClips = int.Parse(firstNode.Attributes.GetNamedItem("clips").Value);
+			manager.allWeapons[i].bulletsPerClip = float.Parse(firstNode.Attributes.GetNamedItem("bulletsPerClip").Value);
+			manager.allWeapons[i].maxClips = float.Parse(firstNode.Attributes.GetNamedItem("clips").Value);
 			manager.allWeapons[i].reloadSpeed = float.Parse(firstNode.Attributes.GetNamedItem("reloadSpeed").Value);
 			manager.allWeapons[i].damage = float.Parse(firstNode.Attributes.GetNamedItem("damage").Value);
 			manager.allWeapons[i].coneAngle = float.Parse(firstNode.Attributes.GetNamedItem("coneAngle").Value);
@@ -65,8 +65,8 @@ public class XMLVendorReader : MonoBehaviour {
 				weapon.range = float.Parse(firstNode.Attributes.GetNamedItem("range").Value);
 				weapon.fireRate = float.Parse(firstNode.Attributes.GetNamedItem("fireRate").Value);
 				weapon.force = float.Parse(firstNode.Attributes.GetNamedItem("force").Value);
-				weapon.bulletsPerClip = int.Parse(firstNode.Attributes.GetNamedItem("bulletsPerClip").Value);
-				weapon.maxClips = int.Parse(firstNode.Attributes.GetNamedItem("clips").Value);
+				weapon.bulletsPerClip = float.Parse(firstNode.Attributes.GetNamedItem("bulletsPerClip").Value);
+				weapon.maxClips = float.Parse(firstNode.Attributes.GetNamedItem("clips").Value);
 				weapon.reloadSpeed = float.Parse(firstNode.Attributes.GetNamedItem("reloadSpeed").Value);
 				weapon.damage = float.Parse(firstNode.Attributes.GetNamedItem("damage").Value);
 				weapon.coneAngle = float.Parse(firstNode.Attributes.GetNamedItem("coneAngle").Value);
